move DerivedIsmConfig to types.ts
