+++ typescript/sdk/src/core/HyperlaneRelayer.ts
@@ -17,9 +17,9 @@ import {
 
 import { DerivedHookConfig, EvmHookReader } from '../hook/EvmHookReader.js';
 import { HookConfigSchema } from '../hook/types.js';
-import { DerivedIsmConfig, EvmIsmReader } from '../ism/EvmIsmReader.js';
+import { EvmIsmReader } from '../ism/EvmIsmReader.js';
 import { BaseMetadataBuilder } from '../ism/metadata/builder.js';
-import { IsmConfigSchema } from '../ism/types.js';
+import { DerivedIsmConfig, IsmConfigSchema } from '../ism/types.js';
 import { MultiProvider } from '../providers/MultiProvider.js';
 import { ChainMap, ChainName } from '../types.js';
