+++ typescript/sdk/src/index.ts
@@ -179,7 +179,7 @@ export {
   ProtocolFeeHookConfig,
   ProtocolFeeSchema,
 } from './hook/types.js';
-export { DerivedIsmConfig, EvmIsmReader } from './ism/EvmIsmReader.js';
+export { EvmIsmReader } from './ism/EvmIsmReader.js';
 export { HyperlaneIsmFactory } from './ism/HyperlaneIsmFactory.js';
 export { BaseMetadataBuilder } from './ism/metadata/builder.js';
 export { decodeIsmMetadata } from './ism/metadata/decode.js';
@@ -195,6 +195,7 @@ export {
   ArbL2ToL1IsmConfigSchema,
   DeployedIsm,
   DeployedIsmType,
+  DerivedIsmConfig,
   DomainRoutingIsmConfig,
   IcaRoutingIsmConfig,
   IsmConfig,
