+++ typescript/sdk/src/ism/metadata/aggregation.ts
@@ -7,8 +7,7 @@ import {
   toHexString,
 } from '@hyperlane-xyz/utils';
 
-import { DerivedIsmConfig } from '../EvmIsmReader.js';
-import { AggregationIsmConfig, IsmType } from '../types.js';
+import { AggregationIsmConfig, DerivedIsmConfig, IsmType } from '../types.js';
 
 import type { BaseMetadataBuilder } from './builder.js';
 import { decodeIsmMetadata } from './decode.js';
