+++ typescript/sdk/src/ism/metadata/routing.ts
@@ -5,8 +5,13 @@ import {
 import { Address, WithAddress, assert } from '@hyperlane-xyz/utils';
 
 import { ChainName } from '../../types.js';
-import { DerivedIsmConfig, EvmIsmReader } from '../EvmIsmReader.js';
-import { DomainRoutingIsmConfig, IsmType, RoutingIsmConfig } from '../types.js';
+import { EvmIsmReader } from '../EvmIsmReader.js';
+import {
+  DerivedIsmConfig,
+  DomainRoutingIsmConfig,
+  IsmType,
+  RoutingIsmConfig,
+} from '../types.js';
 
 import type { BaseMetadataBuilder } from './builder.js';
 import { decodeIsmMetadata } from './decode.js';
