+++ typescript/sdk/src/ism/types.ts
@@ -13,7 +13,12 @@ import {
   TestIsm,
   TrustedRelayerIsm,
 } from '@hyperlane-xyz/core';
-import type { Address, Domain, ValueOf } from '@hyperlane-xyz/utils';
+import type {
+  Address,
+  Domain,
+  ValueOf,
+  WithAddress,
+} from '@hyperlane-xyz/utils';
 
 import { ZHash } from '../metadata/customZodTypes.js';
 import {
@@ -169,6 +174,8 @@ export type AggregationIsmConfig = {
 };
 
 export type IsmConfig = z.infer<typeof IsmConfigSchema>;
+
+export type DerivedIsmConfig = WithAddress<Exclude<IsmConfig, Address>>;
 
 export type DeployedIsmType = {
   [IsmType.CUSTOM]: IInterchainSecurityModule;
